// generated/models/StudentCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StudentCreate
 */
export interface StudentCreate {
    /**
     * 
     * @type {string}
     * @memberof StudentCreate
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof StudentCreate
     */
    groupId: number;
    /**
     * 
     * @type {boolean}
     * @memberof StudentCreate
     */
    status: boolean;
}

/**
 * Check if a given object implements the StudentCreate interface.
 */
export function instanceOfStudentCreate(value: object): value is StudentCreate {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('groupId' in value) || value['groupId'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function StudentCreateFromJSON(json: any): StudentCreate {
    return StudentCreateFromJSONTyped(json, false);
}

export function StudentCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudentCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'groupId': json['groupId'],
        'status': json['status'],
    };
}

export function StudentCreateToJSON(value?: StudentCreate | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'groupId': value['groupId'],
        'status': value['status'],
    };
}